pyproject.toml
[tool.poetry]
name = "save_later"
version = "0.1.0"
description = ""
authors = ["CalebJohn <git@calebjohn.ca>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
feedgenerator = "^2.0.0"
beautifulsoup4 = "^4.11.1"
httpx = "^0.23.0"
uvicorn = "^0.18.2"
readability-lxml = "^0.8.1"
alembic = "^1.8.0"
SQLAlchemy = "^1.4.39"
MarkupSafe = "^2.1.1"
gunicorn = "^20.1.0"
jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pytest = "^7.1.2"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
